+++ Sources/PDKit/Settings/PDLocalSettingsKey.swift
@@ -18,4 +18,5 @@ public enum PDLocalSettingsKey: String, CaseIterable {
     case wipeLocalStoreOnNextLaunch
     case lastCloudKitEventDescription
     case didSeedKVSFromLocal
+    case theme
 }